Separate call APIs from fast-monitor/Form1.cpp
